Fix selection of thread-local wasm when enqueuing
The "onQueueReady" callback was never called when enqueuing to
a shared queue because the code was using vm_id to look up the
thread-local WASM instance in a hash table that was indexed by
vm_key. Now the "on ready" callback is propery fired.

Merging this into envoy-wasm will require a test change there.

Signed-off-by: Gregory Brail <gregbrail@google.com>