Begin to add support for proxy_get_log_level ABI function.
The intent of this function is to return the current log level
in the host. This way, WASM modules can avoid calling proxy_log
at all for certain log levels if they are sure that nothing
will be logged.

This will be important in WASM modules that wish to support "debug"
and "trace" logging but which also want to reduce the cost of
generating log messages that will never be output.

Signed-off-by: Gregory Brail <gregbrail@google.com>