Refactor around VM failure check on Http/Tcp callbacks. (#155)
Refactored around VM failure check on Http and Tcp callbacks in order to handle the VM failure right after it happens. Previously, for example, when panic happens in OnResponseHeaders, then we return Continue since we didn't check the isFail after the Wasm calls. That means Envoy sends the response headers to the client even if the VM fails in OnResponseHeaders, and the failClose is called on OnResponseBody. This seems problematic and unintended.

Relevant to https://github.com/envoyproxy/envoy/issues/14947.

Signed-off-by: Takeshi Yoneda <takeshi@tetrate.io>